+++ packages/swapper/CHANGELOG.md
@@ -1,3 +1,24 @@
+# [@shapeshiftoss/swapper-v9.6.1](https://github.com/shapeshift/lib/compare/@shapeshiftoss/swapper-v9.6.0...@shapeshiftoss/swapper-v9.6.1) (2022-08-04)
+
+
+### Bug Fixes
+
+* fix bch swaps and fix detecting swaps into eth assets ([#945](https://github.com/shapeshift/lib/issues/945)) ([1d9b8d8](https://github.com/shapeshift/lib/commit/1d9b8d8b0cae1ea2bba92f737c8851d43f222c36))
+
+# [@shapeshiftoss/swapper-v9.6.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/swapper-v9.5.0...@shapeshiftoss/swapper-v9.6.0) (2022-08-03)
+
+
+### Features
+
+* detect completed txs and throw error on fail ([#939](https://github.com/shapeshift/lib/issues/939)) ([f85e1fb](https://github.com/shapeshift/lib/commit/f85e1fbbda37da37849a154693a64a5e6c4aa35f))
+
+# [@shapeshiftoss/swapper-v9.5.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/swapper-v9.4.0...@shapeshiftoss/swapper-v9.5.0) (2022-08-02)
+
+
+### Features
+
+* atom thorchain support ([#938](https://github.com/shapeshift/lib/issues/938)) ([86296d1](https://github.com/shapeshift/lib/commit/86296d1ebf3ba57cdc462e1c95e46b122230b881))
+
 # [@shapeshiftoss/swapper-v9.4.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/swapper-v9.3.2...@shapeshiftoss/swapper-v9.4.0) (2022-08-02)